app/serializers.py
from rest_framework import serializers

from app.models import Todo

class TodoSerializer(serializers.ModelSerializer):

    class Meta:
        model = Todo
        fields = '__all__'

    def to_representation(self, instance):
        response = super().to_representation(instance)
        response['user'] = instance.user.username if instance.user else None

        return response